LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-zh-Hant-expected.txt

PASS navigator.language is "zh-tw"
PASS successfullyParsed is true

TEST COMPLETE
